VIH-8946 - created shared method
