+++ VideoWeb/VideoWeb/ClientApp/src/app/waiting-space/consultation-modals/consultation-leave/consultation-leave.component.ts
@@ -1,4 +1,5 @@
 import { AfterViewInit, Component, EventEmitter, Output } from '@angular/core';
+import { ModalTrapFocus } from '../../../shared/modal/modal-trap-focus';
 @Component({
     selector: 'app-consultation-leave',
     templateUrl: './consultation-leave.component.html',
@@ -11,30 +12,7 @@ export class ConsultationLeaveComponent implements AfterViewInit {
     constructor() {}
 
     ngAfterViewInit(): void {
-        // create a trap focus for the modal window
-        const element = document.getElementById('modal-window-confirmation');
-        const focusableEls = element.querySelectorAll(
-            'a[href]:not([disabled]), button:not([disabled]), textarea:not([disabled]), input[type="text"]:not([disabled]), input[type="radio"]:not([disabled]), input[type="checkbox"]:not([disabled]), select:not([disabled])'
-        );
-
-        const firstFocusableEl = focusableEls[0];
-        const lastFocusableEl = focusableEls[focusableEls.length - 1];
-
-        const KEYCODE_TAB = 9;
-
-        element.addEventListener('keydown', function (e) {
-            if (e.key === 'Tab' || e.keyCode === KEYCODE_TAB) {
-                if (e.shiftKey) {
-                    /* shift + tab */ if (document.activeElement === firstFocusableEl) {
-                        e.preventDefault();
-                    }
-                } /* tab */ else {
-                    if (document.activeElement === lastFocusableEl) {
-                        e.preventDefault();
-                    }
-                }
-            }
-        });
+        ModalTrapFocus.trap('modal-window-confirmation');
     }
 
     closeModal() {
